+++ ui/stories/index.js
@@ -3,3 +3,4 @@ import 'src/style/chronograf.scss';
 
 // Kapacitor Stories
 import './kapacitor'
+import './admin'